Add a dedicated response period timer
When receiving a request, the end of the transaction was signaled
by the foreground scan expiration timer. However, it is not expected
to perform a DLL foreground scan during the response period so the
foreground scan expiration timer can't be used for that purpose.
Instead, a dedicated response_period timer is started once a request
is received. When the response_period is expired, the transaction is
signaled as terminated.
In order to prevent the transmission of the response outside the response
period (due to some latency introduced for example by resource congestion)
the D7ASP, D7ANP and DLL layers are also able to check that the response
period is not expired before processing and sending the response packet.