Change the foreground scan policy
- Don't initiate systematically a foreground scan after sending a
packet. Instead, start only the foreground scan if the transmitted
packet is a request (to get the D7ATP acknowledgment or the response).

- According the specification, the responders perform DLL foreground
scan outside the Transaction Periods. So we should wait for the response
period expiration before initiating a new foreground scan if the dialog
is not terminated.

- Don't start a foreground scan when just receiving a foreground frame
during the scan automation period.