Support multiple transaction per dialog
-The responder starts a dialog timeout timer when subsequent requests
are expected after the first request.

-A DLL foreground scan is started by the responder when the transaction
is terminated and is stopped when the dialog timeout timer expires.

-Add the API d7atp_signal_dialog_termination in order to let the upper
layer to decide when a dialog should be terminated.
This way, we prevent d7asp layer to receive new incoming requests while
d7asp state is still in a master state ( d7atp is now considering that the
dialog is terminated only when informed as such by the upper layer).