Support multiple transaction per dialog
-When the DLL foreground scan expires, the dialog is considered as
terminated by the responder

-Add the API d7atp_signal_dialog_termination in order to let the upper
layer to decide when a dialog should be terminated.
This way, we prevent d7asp layer to receive new incoming requests while
d7asp state is still in a master state ( d7atp is now considering that the
dialog is terminated only when informed as such by the upper layer).