Add Frame counter management
The frame counter is incremented each time an encrypted frame is
built. This frame counter is stored in the file system.
A table of trusted nodes is maintained to store the frame counter
contained in the received frames from the peer nodes.
In order to prevent replay attack, the incoming frame counter is
compared against the previous value stored in the corresponding
table entry.