Fix dialog termination after response Tx
When no response period timer and no FG scan timer are running,
the dialog is terminated once the response packet is
transmitted with success or not.

In practice, we can't rely on the fields d7anp_listen_timeout
and packet->d7atp_ctrl.ctrl_tc as they are modified when the
response packet is formatted. Instead, a dedicated flag
"stop_dialog_after_tx" is introduced.

When Tl is equal to 0, the FG scan expiration callback is invoked
immediately after the response period termination.