Add hw_gpio_set_edge_interrupt
This API is used to set specifically the edge on which the
interrupt is fired. This API is likely to be prefered to the API
hw_gpio_configure_interrupt if the edge triggering type
(FALLING/RISING) is regularly modified as this is the case when
enabling the FIFO threshold interrupt.