LoRaWAN: adapted the TimeServer component (which uses RTC timer) to use oss7 scheduler so it can coexist with task scheduled using oss7 timer (which is not using RTC for now).
For now we will continue using RTC for LoRaWAN since the error on the internal oscillator used by the timer currently is too large, especially for the larger intervals used in LoRaWAN (eg 5s for JOIN_DELAY).
This might be revisited later so we can use one timer for everything.