working on power optimations for lrwan1:
lrwan1 platform has a new cmake configuration option: PLATFORM_LOW_CLOCK_SPEED.
When set the clock will be reduced to 4 MHz (instead of 32 MHz) to reduce sleep current (~800 uA).
Note that serial consonlse speed will need to be reduced to 28.8 kbps as well in this case.
This is a temporary setting, until dynamic scaling is implemented