sensor apps: add FS initialisation
FS initialisation can be done by the alp_layer_init function.
But this obliges to start with ALP initilisation prior to the
initilisation of the stack D7A.
However, D7A stack relies on the D7A FS, and not necessarily on ALP.
Thus, it is more safe to invoke explicitely the FS initialisation
in the bootstrap prior to the D7A init.