take hw timer overflow of task time + current hw time into account in timer.c instead of in stm32_common_timer.c
A fix was implemented for a task that would overflow because of the task time + current hw time in stm32 common timer. Some mechanism already existed in the layer above (timer.c). Since the fix in stm32 common timer caused issues when cmp register was set to 0 we adjusted the layer above to account for the "task time + current hw time" as well and removed the previous fix.