D7AP: switch from SLAVE_PENDING_MASTER to MASTER instead of IDLE
A gateway may start sending data while a dialog with one of the nodes is
still active. In this case the D7AP switches to D7ASP_STATE_SLAVE_PENDING_MASTER.

When the dialog in the slave session completes, the master session would
stay pending forever. This in turn would mean the ALP layer would 'leak'
a transfer (completion callback never fires, command would never get
deallocated).

This patch modifies the D7ASP state machine so that D7ASP_STATE_SLAVE_PENDING_MASTER
switches to D7ASP_STATE_MASTER as soon as the slave dialog closes.
So for I have never been able to get a succesfull response to this master transaction,
but at least the ALP layer recovers to a useable state.